Convert $Diff<T, U> -> Omit<T, keyof U>
This is a lossy conversion, since Flow requires that `U` is an object
type and that `T` (1) has all the keys of `U` and (2) the corresponding
values of `T` are assignable to those of `U`. A closer analogue would
use `OmitStrict` instead of `Omit` (where
`type OmitStrict<T, K extends keyof T> = Pick<T, Exclude<keyof T, K>>`)
but using the standard `Omit` means we don't need to add
potentially-conflicting definitions elsewhere to import here.